used async method to execute
This is needed since, system(“cmd”) will block, and when it is
blocking, if user types, user will directly type on the status bar. If
user enters <Enter>, it may even cause status line to grow. Which is
almost unrecoverable. Using async execution will make it block as
little as possible.