fixed a bug that causes delphi to execute more than necessary
Originally it uses the builtin &modified variable to keep track of if
the file is modified and thus need to execute again. This works fine if
we save the file immediately after we make every change like we did
before. But Delphi does not save the file automatically anymore, so we
needs to keep track of the dirty by ourself.